implemented creation of report and faclities search

+++ ISA - project/src/main/java/com/javaguide/ISAprojekat/service/UserService.java
@@ -38,7 +38,7 @@ public class UserService {
         return clientRepository.findByEmail(email);
     }
 
-    public Client findOne(Long id) {
+    public Client findOne(Integer id) {
         return clientRepository.findById(id).orElseGet(null);
 
     }
